Change the default card Sched/LiveTVOrder to 1.
The current input configuration code can't easily mimic the old
behavior of using cardid as the default.  It will be changed back
eventually, but for now, use 1 instead of 0 as the default.  It's easy
to forget to change when configuring a new card and leaving it at 0
causes the card to be ignored.
changes. Lines starting # with '#' will be ignored, and an empty
message aborts the commit.  # On branch inputs # Changes to be
committed: # modified: bindings/perl/MythTV.pm # modified:
bindings/python/MythTV/static.py # modified:
libs/libmythbase/mythversion.h # modified: libs/libmythtv/cardutil.cpp
bindings/perl/MYMETA.json #