MythTranscode: Reset the video stream ID before generating keyframe index
The ID of the video stream in the transcoded file may not be the
same as that in the input file.  Generally this is the case
and most transcodes generate a correct keyframe index.  However,
some channels (e.g. UK Freeview horror channel) have audio as the
first channel in the input file.  In this case, lossless transcode
makes the first stream video and so the keyframe index generation
is incorrect.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>