HTTP: Split the frontend status page out of the code and into html/
This results in a temporary loss of information from this page such as
the list of backends. This will be re-added once the information is
made available through the services API.

Some new information has been added, but it's not well presented
currently. The Frontend service of the API needs some serious work so
we can make a proper job of it.