Check the title instead of cardid when a slave backend connects.
Previously, a valid cardid with a zero input was used to signify a
free card when a slave connects.  The card/input merge changed the
zero inputid check to use cardid which is wrong since the cardid
should always be valid.  Check for an empty title now to signify a
free card.