Initial websocket implementation
For now it operates on it's own port (6545) until I'm able to refactor
the HTTP code and allow it to share the status/ssl ports (6544/6554).

I need to a plugin extension capability to actually do anything useful
with it. For now it just spits out internal events on the wire, which
was the primary motivation for adding the feature. The WebFrontend
will monitor events and update pages in realtime to reflect events
such as schedule changes, recording addition/deletion, status changes
etc with no refreshing or polling.

Third Party application developers please note that since these are
internal events their stability (name, behaviour and existence) is not
guarenteed. This is not part of the Services API, so use them at your
own risk.