WebSocket: Split the MythEvent functionality into an extension
Adds support for server-side filtering of events, so clients only
receive those that they are interested in. These are specified by the
client in a text frame with the format:
WS_EVENT_SET_FILTER {EVENT_NAME} {EVENT_NAME2} {EVENT_NAME3}

A special value of "ALL" in place of the event names will enabled
sending of all events to the client. It's recommended that the client
is more specific about their needs as a backend may in theory send any
number of events per second, most of which the client will have no
interest in. For a mobile client this may even result in unexpected
high data usage!

It also changes the behaviour so that events are not sent by default
and they can be enabled or disabled on demand without affecting
filters or closing the socket.

To enable sending of events:
WS_EVENT_ENABLE

To disable:
WS_EVENT_DISABLE