Fix the value of the starting channel when no channels have been defined.
The setting code assumes it should use the label in the event that the
value is empty. This resulted in the starting channel being set to
"Please add"

Since this code will hopefully disappear eventually it's not worth trying to change the assumption since that might break other settings which rely on the behaviour. An invalid starting channel will get ignored ultimately.