Fix DVB-S2 tuning
I'm baffled by this, but DVB-S2 tuning which was working perfectly for
years until 48 hours ago suddenly started to fail for all multiplexes.

Reviewing the code it appeared some code to treat DVB-S2 channels as
DVB-S2 instead of DVB-S was disabled. Enabling it has allowed S2
channels to be tuned again. It's unlikely that I'm the only person
affected, so let's make this change for everyone.