Services: Split GetSetting into GetSettingList, and GetSetting which only returns a single value.
GetSetting now returns a single string containing either the value of
the requested setting, or if it cannot be found, the default value if supplied.

The existing behaviour made it harder to use GetSetting than it needed
to be for it's most common purpose which is accessing a single
setting. Although I've retained the ability to return all settings on
a host in the new GetSettingList, it's difficult to imagine a scenario
where it would be of value.