Allow some record table columns to be NULL.
Previously, the invalid value "0000-00-00 00:00:00" was used for
MythDates that were not applicable.  With Qt5, invalid values get
translated to NULL when saving to the database and this causes an
error if the column is not configured to allow NULLs.  This commit
changes the affected columns in the record table and the supporting
code to allow NULLs.