Add skeletal user login and session management class
This is only currently used to manage logins for the WebFrontend and
it isn't a complete implementation but it could be extended in future
to support multi-user features.

There is no UI yet to add/remove users, set passwords or permissions.

The design is session based, with the master backend acting as the
central, and only, manager for sessions. Users log in on any client,
but the actual request is processed on the master. A successful login
results in the creation of a 'session' identified with a session token
which the client will use in all subsequent exchanges to minimise the
login credentials going over the network. Sessions have expiry dates
and can be terminated at the master backend effectively allowing users
to be 'logged out' remotely.

For the WebFrontend sessions end when the browser is closed, unless
TLS is being used in which case a cookie is set which allows the
session to be resumed.

There is a stub for a context-dependent permissions system, extensible
by plugins. Other multi-user features such as restricting usage for
certain users to a set period during the day, or a fixed number of
hours can be bolted on later.