HTTP: Add partial handling of OPTIONS and framework for 405 errors
This doesn't entirely work because some of the server extensions are
ambigious, we can't know which ones apply without actually trying to
handle the request, which isn't optimal when we only want to return a
two-line header. On possibility is to eliminate the ambiguity, move
the upnp code to it's own sub-url e.g. /upnp/.

A possible upside to proper 405/OPTIONS support is that the services
code could indicate whether GET or POST are allowed for each endpoint
using HTTP protocol as it was intended. There are some who advocate
for proper use of OPTIONS with REST apis and it costs us nothing to
offer this capability.