Partial fix for crash with editor under QT5. Refs #12394
There are other places in this code where we modify the map while
iterating over it, some of those require a very small amount of
refactoring to eliminate. I just haven't found the time to fix them so
I'm pushing what I've already done in the hope that it reduces the
number of crashes in the short term.