WebFrontend: Limit length of the programme description in the guide
Uses a very rough system, based on the available space and a fixed
upper limit of 600 characters to reduce loading times. It's not
perfect, but this reduces the amount of data we need to send over what
might be a slow network connection.

It can be tweaked later if the results aren't suitable, e.g. for those
using a smaller font size than the default.

This isn't the only optimisation that will be made, it's just one of
the easiest ones to start with.