Services API: Fix a logical issue with GetProgramGuide
GetProgramGuide took an argument - StartChanId but since the list of
channels is sorted by the Channel Number, and not it's ID this made no
sense. This meant there was no way of restricting the results to just
a subset (page) of the guide.

So to address this issue, and make GetProgramGuide consistent with
other service API methods StartChanId has been replaced with
StartIndex, and NumChannels has been replaced with Count.

'TotalAvailable' is included in the results, but for now it will only
contain zero. A fix for this will be committed soon, I hope.

Unfortunately, the changes to make this necessary have negated the
earlier performance improvements. Fetching the complete guide is now
25-50% slower than it was before this commit, this is obviously not
acceptable and I'm trying a variety of approaches to get performance
back at least to where it was. On the other hand, since data can now
be fetched in ranges, perceived loading time can be reduced by
backgrounding the loading of most of the guide.