Subtitles: Fix Qt5-related glitches in the black background.
At least I assume this came from the Qt5 update.

The black background that is drawn behind chunks of text started being
rendered smaller by what looks like 1 pixel.  When a line of text is
drawn as multiple adjacent chunks (due to changes in formatting or
color), a small gap would show up between each chunk.

The solution seems to be to define a pen style with width 0 and the
same color as the background, instead of using the Qt::NoPen default.