Subtitles: Fix automatic display of captions/subtitles.
There was something of a race condition in the previous code.  At
startup of playback, the decoder thread runs ahead and updates track
lists as it encounters them, including audio, cc608, cc708, etc.  If
no subtitle tracks are found before the UI thread gets to a certain
point, the UI thread commits to keeping subtitles disabled, regardless
of the "automatically display subtitles" setting.

In the new version, at startup of playback, the UI thread keeps the
original intention alive so that subtitles will automatically start
displaying when they eventually arrive.

Note that this fix doesn't address the problem where the player ought
to switch to a "better" subtitle track type when it comes along, such
as switching from cc608 to cc708.

Refs #10740 (only because this fixes a fix with that ref)