MythVideo: Add a log message describing mysterious behavior.
MythVideo delays destroying its data for 3 seconds after exiting, in
case the user exited by mistake and reenters within 3 seconds.  This
can be confusing if the user was exiting/reentering in order to effect
a file system rescan.

It's probably a bit too intrusive to deliver a pop-up notfication in
the UI when the data is reused, but at least we can describe it in a
log message.