Add utility marks for preroll amount and latest playback position.
==== Program start:

The recorder writes a mark of the new type MARK_UTIL_PROGSTART as soon
as it encounters a keyframe close to the recording's scheduled start
time.  This essentially skips past the preroll region and the "start
early" region.

The frontend starts playback from this keyframe if the user selects
Play and there is no explicit bookmark (but not if the user selects
one of the "Play from..." menu items).

==== Playback position:

During playback, a timer event auto-saves the current playback frame
with the new MARK_UTIL_LASTPLAYPOS mark type, every 30 seconds. The
actual saving is done in a background thread to avoid playback
glitches.  Saving is done only during active playback, i.e. not while
paused.  The MARK_UTIL_LASTPLAYPOS mark is tidied up (cleared) if the
user saves a bookmark while exiting playback.

The "Play from..." menu gets a new item (a new translatable string):
    Play from last played position

This item is added to the menu only when the recording markup actually
contains a MARK_UTIL_LASTPLAYPOS mark.

The motivation of MARK_UTIL_LASTPLAYPOS is that if you accidentally
exit playback without setting a bookmark (or the frontend crashes, or
the OS crashes, or ...), this allows you to resume roughly where you
left off.  Because of the 30-second timer, if you start playing
normally and realize you've made a mistake, you have 30 seconds to
exit playback and instead "Play from last played position".

Note that different frontends can fight over setting this mark, same
as with a regular bookmark.

Still to do:
1. Preview generator can do less guessing about preroll.
2. Provide MythVideo launcher with similar progstart/lastplaypos options.

Refs #11713.