Be less aggressive about clearing the MARK_UTIL_LASTPLAYPOS mark.
Refs 11713.  We actually don't want to tidy up the mark when then user
saves a bookmark at exit.  If the user starts playback only to realize
they should have started at the last played position, we want to give
them the chance to exit playback - possibly setting a bookmark but not
clearing the last played position - and restart at the last played
position.

On the other hand, it is still OK to tidy it when the bookmark is
cleared when playback reaches the end of the recording.