Turn off LastPlayPos loading by default.  Refs #11713.
In the way things were set up originally, by default any playback
would check for bookmark, progstart mark, or lastplaypos mark, and
start playback from there, unless any or all of these had been
explicitly disallowed.  This was fine for playback from the Watch
Recordings screen, which would always explicitly set each permission.

It was not so good for other forms of playback, such as from
MythVideo, mythavtest, or Gallery, since they would all favor
lastplaypos, which is periodically updated for all playback.

As a result, lastplaypos is changed to be disallowed by default.

This is not really relevant to the progstart mark, since it is only
present for recordings (though it could become a minor issue if a
recording and its metadata are imported into MythVideo).