Fix lirc being ignored after dbus screen saver changes.
If the Screensaver is Asleep() we ignore all keypresses
generated via the remote. This is clearly so that a button
pressed on the remote while the screen is asleep only wakes
up the screen, rather than triggering an action.

This swaps the logic around in the dbus screensaver so that
it satisfies the Asleep() check.

We may need to revist this in the future if we find that
when using the dbus screensaver we are not ignoring the
remote keypress which is used to bring us out of screensave.

Fixes #12435
Refs #12414