Subtitles: Fix cc608 preamble indents.  Fixes #12051.
Cherry-picked from 366c428708682420bbf53a301d145ca441aedbdd .

Thanks to Helen (faginbagin) for the analysis and patch.

As Helen points out in the ticket, it would be all-around better if we
took an approach like the cc708 captions and rendered everything onto
an internal grid of characters and then converted that to strings to
draw on the screen, but that's a lot of work for a mostly deprecated
caption format.