Subtitles: Fix incorrect rendering of non-ASCII cc608 & teletext chars.
Special characters in the non-ASCII range were being displayed
incorrectly in cc608 and teletext captions.  This was due to sign
extension issues, and is fixed by converting it to QLatin1Char before
converting to QChar.  This is likely just an issue for Qt5.

In mythccextractor, the language-specific tables for teletext special
character mapping were in the wrong order, such that German and French
were swapped, and Italian and Czech/Slovak were swapped.  This was
just a matter of reading the bit order in the spec in reverse.