Add support for building the BD-J support classes in libmythbluray.  They are not used yet but building them is the first step towards using them :)
This feature can be enabled by running configure with the '--enable-bdjava' option.

To keep the build environment as close to upstream as possible, the 'ant' build tool must be installed and in the path.  As this feature builds Java classes, a JDK must of course also be installed.  The JDK_HOME environment variable can be used to indicate where the JDK is located, or passed as a parameter to configure.