Change vector excluded_card[id]s to scalar excluded_input.
Several methods passed a vector of cardids to specify to specify which
should be ignored.  There was never more than one cardid in the list,
so we can get by with a slightly simpler API using just a scalar.
Since input is now the preferred term instead of card, make that
change too.