Add new GET_FREE_INPUT_INFO protocol command.
The intent of this change is multi-faceted.  The logic is now much
simpler than the myriad other commands that already existed.  The old
commands and their support will be removed in following changes.  It
also provides more information to the frontend so it can make more
intelligent decisions.  The appllicable tvremoteutil methods are
re-implemented using the new command.  Following commits will attempt
simplify the logic using them.