Allow mythccextractor to be run on a recording which is still in progress.
Pull the program info in from the DB so the ringbuffer can tell if it is
being written to.  IsRegisteredFileForWrite does not get set right away so
set a flag to wait for it, if the recording is still in progress.
    
When processing an in-progress recording, don't show the percentage done,
since it is meaningless.