Add VBox TV Gateway support
VBox is a networked tuner similar to the HDHomerun devices and is available
in 2 and 4 tuner versions with DVB T/T2, DVB S/S2 and DVB-C tuners.

VBox support is compiled in by default since it doesn't require any external
dependencies but can be disabled by passing --disable-vbox to configure if
required.

See the wiki here www.mythtv.org/wiki/VBox for more details.

Note: Bumps the database version to 1340.