VBox: add channel filtering to the channel scanner
This makes sure we honour the FTA and TV/TV+RADIO options in the channel
scanner when using the 'VBox Channel Import' option.

This also adds support to the channel scanner for the 4 tuner vboxes with
mixed DVB-T/T2/S/S2 tuners ensuring only channels supported by the selected
tuner are added to the video source.

Currently there is no way in the VBox API to determine if a channel requires a
T or T2 tuner or S or S2 tuner. The Vbox devs are aware of this problem and
are working to fix it in the next firmware version.