mythutil: add a new --findlyrics command
This will use any compatible scripts found in
{SHAREDIR}/mythtv/metadata/Music/lyrics/ to search for lyrics for a given
MythMusic songid.

There are some simple example scripts that will search for lyrics in a file,
embedded in a track's tag or will also use any public APIs provided by some
lyrics web sites. Each script has a priority that will determine the order
the scripts will be tried. See the README for the options the scripts need to
support and the format they should return.

Most of the scripts are taken from a CU LRC Lyrics kodi plugin but have been
modified for our purposes.

This is just the first step to add lyrics support to MythMusic.