RemoteFile::CopyFile: Don't overwrite existing files by default
CopyFile doesn't check for a file's existance before writing to it.
For storage groups containing multiple dirs this can lead to multiple versions of a remote file (only
one of which will be read).

This patch will fail the copy operation (by default) if the destination file already exists.
If the overwrite flag is set the destination file is deleted first, ensuring that multiple versions
are never created.

It also adds optional verification that the copy succeeded (by comparing file sizes) as writing can fail
silently if the filesystem fills up.

Refs #12424