UI: Avoid texture memory exhaustion by shrinking the painter image cache
This dynamically shrinks the OpenGL painter's m_MaxHardwareCacheSize
when a texture can't be created.

This also fixes a bug in MythRenderOpenGL::ClearTexture which wasn't
reporting the OOM failure from glTexImage2D.

The Raspberry Pi has a limited amount of GPU memory (typically 256MB)
and the texture buffers overflow if the painter hardware cache size
is set at the current 96MB.  This patch adds independent configurable
sizes for the hardware and software caches (UIPainterMaxCache[HW|SW])

The patch also reduces the default hardware cache size to 64MB and
adds 64 bit correctness to the SetMaximumCacheSizes function.

NB the texture buffer overflow is most noticeable on the MythCenter
Xmas theme which uses an animated background.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>