Attempt to fix configuration of diseqc inputs.
This was broken when cards and inputs were merged.  Previously,
logical inputs were added to an existing, DVB-S card as needed.  Now,
the user must explicitly add a new card using the same device for each
additional input.

Note that the current, input configuration is less than ideal.  It
still keeps the idea of cards and inputs alive even though card and
input are now synonymous.  This will all be fixed in time.

Refs #12499