Fiddle with the theme template regular expression for #12003
It appears that \w worked for a long time. But does not match #
so it was changed to \S which matches everything but white space.

Now try with \w and explicit #.

The regexp is tested with online tools.
(I couldn't get QRegExp-Editor to build and refactoring to suit unit
tests didn't look easy)

I'll pull it into fixes/0.27 once I hear success reports.

Refs #12003