TV: Avoid a race condition starting the decoder thread
MythPlayer::StartPlaying creates and starts the decoder in the paused state.
It then performs some housekeeping and unpauses the decoder.  However,
if the decoder thread hasn't run during this period then the unpause
will have no effect and when eventually DecoderLoop runs it will
immediately pause the decoder and never decode any frames.

This bug manifests itself on the RPi when starting PiP recordings,
especially of MP4 recordings.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>