configure: Link with the correct OpenGL/ES library
Currently configure adds the library -lGL if the basic OpenGL headers
and libraries are detected.  Later, it checks whether Qt was built with
OpenGLES support and adds those libs.  This causes problems at runtime
when using Qt's EGLFS graphics which attempt to use shader programs
(part of ES 2.0) which then fail because the basic OpenGL library
is being referenced.

This patch removes the basic OpenGL lib and replaces it with
the version being used by Qt.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>