TV: OpenMAX compatibility with the Bellagio library implementation
This patch permits building the OpenMAX code with the Bellagio
library (Ubuntu >= 12.04).

Allows reduced functionality (no h/w deinterlace) if image_fx
component is missing.

NB the v0.9.3 Bellagio implementation on Ubuntu 12.04 has these
limitations:

- No video decoder component
- No image_fx component, therefore no deinterlace
- The video renderers (fbdev_sink & xvideosink) time out when
  disabling the input port during initialisation.
- The alsasink audio render times out on transit to the idle state.

Although the code compiles (and works on a RPi) the Bellagio
implementation doesn't appear to be able to send callbacks and is
thus unusable.  This maybe a pthread library incompatibility.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>