Force OpenGL1 line width to be at least 1.
The glLineWidth() call requires its argument to be at least 1.  If we
pass 0, OpenGL sets an error code that may be detected later via
glGetError(), and misattributed to some other situation.

This was causing the first chunk of caption text to be blank because
the black background rectangles were being drawn with line width 0.