TV: Qt5 compatibility for DBEvent::GetMatch
Qt5 changed the return type of QDateTime::secsTo() to qint64
from int.  As a result std::abs of this type requires a conversion
which is ambiguous (int, float, double) and gives errors with
C++11 compliant compilers.

This patch uses the qAbs function which is Qt4 and 5 compatible.

Thanks to Craig Treleaven <ctreleaven@cogeco.ca> for spotting this.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>