MythGallery: Random slideshow with a seasonal theme
This creates a new mode of slideshow, called "Seasonal," to supplement the existing sequential and random slideshows.
It shows a random slideshow but with a larger weight given to photos from the same time of year.
The feature is only enabled if EXIF support is present.

I also made some fixes to the sequence classes, so that they tracked history better and could run forever without leaking memory.

Fixes #12335

Signed-off-by: Roger Siddons <rsiddons@mythtv.org>