Get around libvdpau crashing with some resolutions.
Two problems at hand, one is ours being that the pitch for U plane was wrong (twice as big as it should have been).
There's still appear to be a problem with libvdpau causing an out of bound read when using resolution not multiple of 16. The issue is bypassed here as we allocate more than necessary and have both the U and V plans pointing to the same buffer, giving us plenty of leeway.

Fixes #12531