MythGallery: Fix cppcheck warnings
Prevent:

Class ' GalleryFilter' has a constructor with 1 argument that is not explicit.
Class ' ChildCountEvent' has a constructor with 1 argument that is not explicit.
Class ' ChildCountThread' has a constructor with 1 argument that is not explicit.
Class ' ImportThread' has a constructor with 1 argument that is not explicit.
Class ' ThumbGenEvent' has a constructor with 1 argument that is not explicit.
Class ' LoadAlbumListener' has a constructor with 1 argument that is not explicit.
' class ImageView' does not have a copy constructor which is recommended since the class contains a pointer to allocated memory.