Fix the update to schema version 1337.
Non-device cards were incorrectly grouped together in some cases.
While it is possible that multirec was really intended for those
cases, it's more likely that it wasn't.  Note this fix will not do
anything for those who have already upgraded.  For them, the dameage
has already been done and needs to be fixed manually.

Fixes #12605