extend character encoding fixups for Cyfra+ and Polsat
This extends the framework to allow applying fixups based on the
transport that the data is carried on, instead of the transport that the
data is about.
It is now possible to remove or replace character encoding signalling.
The old code would only add, but not replace the signalling.

Based on a patch by Wagabunda http://lists.mythtv.org/pipermail/mythtv-dev/2008-December/064048.html
Updated by Voyo and Piotr Oniszczuk

Fixes #9480