MythNetvision: Add page navigation
This adds the possibility to navigate to the next/previous page and skip 10
pages at a time.  Trying to skip 10 pages for a grabber that does not support
skipping to a specific page (currently only the YouTube grabber) only skips one
page.

Based on patch by Angela Schmid <angela.schmid@wolke7.net>.

Fixes #12071