add missing space between sentences of DiSEqC command repeat hack
Rationale for space vs. newline is that the info widget is three
lines on all tested themes in the connection dialog.

The whole function should be changed two use separate repeat count
and repeat mode settings, as the main rationale for this interesting
behaviour was to avoid a schema change.