MythMusic: update the list of available streaming radio services
This removes the static list shipped with Myth and replaces it with a list
downloaded from our server. This makes it possible for users to get an up
to date list without having to update Myth.

The list has grown from the few hundred streams we had previously to over
31000 from all over the world. You can now filter the list by country and
language as well as broadcaster, genre and channel.

We're currently using a slightly sanitized version of the user updated
database available from http://www.radiosure.com/stations/ so if you spot a
problem or need to add a new stream please help everyone by updating the
details there.

THEMERS NOTE:  there have been some changes to the 'searchstream' and
'editstreammetadata' screens

NOTE: bumps the ABI version so both core and plugins need to be recompiled.