The port numbers for the websocket (for the backend webserver), and the upnp ssl port were hardcoded.
This changes the default port number to be
- backendport + 5  (websocket port)
- backendport + 10 (upnp ssl port)

This leaves the default upnp ssl port unchanged, and moves
the websocket port to one which does not conflict.

This allows 2 instances of the backend to run simultaneously
on the same system (ie. dev & prod), provided that the backend
port for one of them is changed to a different one.