Refs #12307 Check if UDisks2 is in use.
First check for the original UDisks service,
if we fail to find it, check for UDisks2.

If we find UDisks2 service, then this system
has moved to UDisks2, rather than UDisks1.

For now log a message and give up. This will
reduce the frontend startup time by the 5 seconds
the loop takes to complete and fail.

TODO: Implement the same support via UDisks2 interface