Restore queuing of jobs when a recording finishes.
Reverts a portion of ffef83f. In TVRec::FinishedRecording, a check
of the recording's filesize is made. If it's less than 1000, then
PreviewGeneratorQueue isn't run. JOB_COMMFLAG and JOB_TRANSCODE
get removed from the queue.

After ffef83f ProgramInfo::GetFilesize will, at times, return a
filesize of 0 even though it's not. This is always 0 at the end
of a recording (only DVB and HDHOMERUN card types tested.)

This workaround will use the RecordedInfo filesize value if it's
non zero because in all tests, if non zero, the RI filesize
matched the value returned from the DB.

If RI filesize is zero, then there's no way to know if it's valid,
so use the 'old' DB query. If VB_RECORD is set and the DB filesize
is > 0 (and RI filesize = 0) then LOG.

I'm assuming that simply removing the 1000 byte test is a worse
workaround to the problem.

Refs #12290

(cherry picked from commit 2e98fb583276c3966fed21fab354b11c7e5d591c)