Pi: Background shows through with ffmpeg & openmax
aa676f77 prevented background bleed where the decoder was openmax
but does not cover the case where the decoder is ffmpeg and the renderer is openmax,
which still bleeds through. This is a viable combination which provides better quality
than openmax decoder in some cases.

Also fixes a segfault documented by #12700 & #12709, which was introduced by the fix for #12691.

Fixes #12693

Signed-off-by: Roger Siddons <rsiddons@mythtv.org>