Python: Restore chanId/startTime API for recordings
Exisiting python scripts depend on the Recorded class taking the chanid/starttime instead of the recordedid as the primary key.
This patch updates the Recorded class to continue using the chanid/starttime pair.
A future patch can add recordedid as a second option so as not to break backwards compatibility.

Fixes: #12260