HTTP Server: Provide additional CORS support
This commit adds a new AllowedOriginsList setting that
stores a comma separated list of Origins. If the setting
is missing, the list defaults to:

    https://chromecast.mythtv.org
    http://chromecast.mythtvcast.com

Access-Control-Allow -Credentials and -Headers were also
added to the response to support Chromecast.

A TODO item is closed by adding the hostname to
the existing list of IP addresses (no change in
the ports checked.)

Note that previously, if ResponseTypeFile was found, the
CORS header wasn't being added. This caused the Chromecast
feature to fail.

There is no GUI. Clients have the ability to change the value
with the Services API. May need to rethink using settings as
this is a security risk.

For example, it a valid Origin exists in the request, the
following will appear in the response:

    Access-Control-Allow-Origin: https://example.com
    Access-Control-Allow-Credentials: true
    Access-Control-Allow-Headers: Content-Type

If there is no match, an existing critical log will fire:

    Cross-origin request received with origin (https://example.net)

Missing http://, https:// or a '*' character in list entries will
be ignored, and the following will be logged:

    Illegal AllowedOriginsList entry '*'. Must start with http[s]:// and not be *

(cherry picked from commit 5292483764357d86f6dc2b155b262c2c072748da)