HTTP Server: Provide CORS (Chromecast) support
This commit adds a new AllowedOriginsList setting that
stores a comma separated list of Origins. If the setting
is missing, the list defaults to:

    https://chromecast.mythtv.org
    http://chromecast.mythtvcast.com

Access-Control-Allow -Credentials and -Headers were also
added to the response to support Chromecast.

There is no GUI. Clients have the ability to change the value
with the Services API. May need to rethink using settings as
this is a security risk.

For example, if a valid Origin exists in the request, the
following will appear in the response:

    Access-Control-Allow-Origin: https://example.com
    Access-Control-Allow-Credentials: true
    Access-Control-Allow-Headers: Content-Type

If there is no match, an existing critical log will fire:

    Cross-origin request received with origin (https://example.net)

Missing http://, https:// or a '*' character in list entries will
be ignored, and the following will be logged:

    Illegal AllowedOriginsList entry '*'. Must start with http[s]:// and not be *

Modified from the 0.28/29.0-pre version for 0.27