Change the scheduler to honor reclimit.
Thie commit along with earlier ones involving reclimit and reccount
now make it possible to schedule more than one recording or live TV
session at a time on a single tuners that allow it(*).  Think of this
as multirec without additioinal virtual tuners.  Note that this is not
a magic bullet and the multiplex and channel restrictions that
multirec imposes still apply here.  The advantage is that it hides the
virtual tuner complexity from the user and makes some scheduling cases
easier and faster.

(*)None of the recorders actually support this yet.  That will start
in the hopefully not too distant future.