Add the remaining support for the Schedule as Group feature.
When enabled for an device, all virtual inputs for it will be
scheduled as a single, grouped input.  Programs will not be assigned
to a specific, virtual input until right before recording is to start.
This is much more efficient than scheduling each virtual input
individually.  While this doesn't reduce the scheduler's big O from
O(N^2), or possibly even O(N^3), it can greately reduce N.

The current down side with enabling Schedule as Group, is that the
scheduler assumes the grouped input can handle an unlimited number of
simultaneous recordings.  In reality, the input still has to honor its
maximum recording setting because MythTV requires all such resources
to be pre-allocated.  Consequently, it is possible to run out of
inputs and have reocrdings fail in worst case situations.  Therefore,
be sure to overprovision the maximum recordings setting for your
expected situation.