Work around a suspected Qt bug affecting some settings.
There is a suspected bug in some versions of Qt, notably in version
5.6.1 in Debian, where calling QVariant::toString() returns garbage if
the MySQL type is TINYINT.  Because MythTV does this for all settings,
it is possible for a setting to lose its previous value when a
settings dialog is entered.  The work around is to have MySQL cast the
value to a string type before it gets to MythTV.  Thanks to Karl Dietz
for suggesting this fix.

(cherry picked from commit b113943461bb26dd411bfe625082fab44b1390ed)