Serialization Implementation for XML: gcc6/Qt5.{6,7} SEGVs
This isn't a MythTV bug, rather an issue with Qt5.{6,7,??}
and gcc version 6. This will get Arch users with the above
combinations working until a proper solution is developed.

In XmlSerializer::GetContentName when pMetaObject is NULL,
a SEGV fires rather that getting a -1 return from
pMetaObject->indexOfClassInfo().

Examples: <be>:6544/Myth/version and <be>:6544/Myth/GetHostName.
Both fail on Arch linux with the above combinations.

Thanks for the research on: bbs.archlinux.org/viewtopic.php?pid=1631996,
testing on the ticket, and the reference to:
http://lists.qt-project.org/pipermail/development/2016-March/025153.html

Refs #12782

(cherry picked from commit 38de43efa92603e2ac524b9a72951c15caf2af4e)