Fix Seg Fault due to OSD painter being deleted while image loader threads are still using it.
Image loading for the OSD uses ImageLoadThread and continues
loading while the video is playing. If the video playback
encounters an "InputChanged" situation it deletes the OSD and its
painter and re-creates it.  If the Image load is still running and
using the painter, it gets a Seg Fault calling a virtual function
using a corrupted object

Exception Details:
1   in MythPainter::GetFormatImage at mythpainter.cpp:540

Fix: Keep the old osd painter valid until either it changes again
or the OSD is deleted.

Refs #8096