Fix Seg fault in OSD due to notification being displayed using a painter that is no longer valid.
The OSD goes through all notifications in the notification center
and displays them. Each notification screen has a painter stored in
it. The painter will be invalid if the notification started being
displayed before the playback started or before an inputchanged
notification caused it to be deleted

Exception Details:
1   in MythUIType::HandleMovementPulse at mythuitype.cpp:374
...
3   in OSD::DrawDirect at osd.cpp:702

Fix: Always update the painter to the latest one before painting
notifications.

Refs #8096