Frontend and backend changes to support music choice.
Music choice channels from Comcast are mpeg2 video which internally
identifies as 30 fps but actually sends frames with intervals
5.939, 0.033, 0.033, repeat, ... This amounts to 3 frames at 30fps
every 6 seconds. These frames are never identified as key frames.
Tuning takes up to 8 seconds for these channels on CETON. This
stream breaks several MythTV timeouts, and causes havoc with the
audio-video synchronization routines. It causes the recorder to
report a great many 6 second gaps in the video and mark it as failed.

Fix: Increase timeouts. Increase audio buffer size to allow up to
8 seconds to be buffered. In the backend allow recordings to start with a
non-key frame and set a high threshold for gaps in specific cases.
Change handling of video buffer waits so that audio data is not lost by
audio resets in music choice programs.

Fixes #8096