Fix FIFO transcoding.
There were two issues at play.
One that was a regression over 0.27, in that it was assumed the stride was identical to the width which is no longer true from 0.28
Second, if the file being transcoded changed resolution, the video output would have been corrupted.

TODO: Handle codecs not returning a YUV420 frame.

Fixes #12479